+++ src/render/extract.rs
@@ -5,7 +5,7 @@ use bevy::{
 
 use crate::light::{AmbientLight2d, PointLight2d};
 
-#[derive(Component, Default, Clone, ShaderType, Copy)]
+#[derive(Component, Default, Clone, ShaderType)]
 pub struct ExtractedPointLight2d {
     pub transform: Vec2,
     pub radius: f32,
@@ -35,6 +35,16 @@ pub fn extract_point_lights(
             falloff: point_light.falloff,
         });
     }
+
+    // BufferVec won't write to the GPU if there aren't any point lights.
+    // For now we can spawn an empty point light to get around this.
+    commands.spawn(ExtractedPointLight2d {
+        transform: Vec2::ZERO,
+        intensity: 0.0,
+        radius: 0.0,
+        falloff: 0.0,
+        color: LinearRgba::BLACK,
+    });
 }
 
 pub fn extract_ambient_lights(